\name{teapot}
\docType{data}
\alias{teapot}
\title{Utah Teapot}
\description{
  The Utah teapot is a classic computer graphics example.  This data set
  contains a representation in terms of triangles.
}
\usage{
data(teapot)
}
\format{
  A list with components \code{vertices} and
  \code{edges}. \code{vertices} is a 3 by 1976 numeric matrix of the
  coordinates of the vertices. \code{edges} is a 3 by 3751 integer matrix
  of the  indices of the triangles.}
\source{
  Converted from the netCDF file that was at one time made available by Dave Forrest at
  \verb{http://www.maplepark.com/~drf5n/extras/teapot.nc}.
}
\keyword{datasets}
